Modularized application: added the module-info.java, the necessary plugins and split the project into three modules
