+++ app/src/main/java/calendar/backend/service/application/CalendarBackendServiceApplication.java
@@ -1,0 +1,13 @@
+package calendar.backend.service.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CalendarBackendServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CalendarBackendServiceApplication.class, args);
+	}
+
+}
