+++ app/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module app {
+    requires spring.boot;
+    requires spring.boot.autoconfigure;
+    requires service;
+    requires data;
+}
