+++ data/src/main/java/calendar/backend/service/data/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package calendar.backend.service.data.repository;
+
+import calendar.backend.service.data.model.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, String> {
+}
