+++ data/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module data {
+    requires spring.context;
+    requires spring.data.commons;
+    requires spring.data.mongodb;
+    requires static lombok;
+
+    exports calendar.backend.service.data.model;
+    exports calendar.backend.service.data.repository;
+}
