+++ service/src/main/resources/application-dev.yml
@@ -1,0 +1,6 @@
+spring:
+  data:
+    mongodb:
+      uri: mongodb://localhost:27017
+      database: admin
+      use-ssl: false
