--- src/main/java/calendar/backend/service/calendarbackendservice/CalendarBackendServiceApplication.java
@@ -1,13 +1,0 @@
-package calendar.backend.service.calendarbackendservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CalendarBackendServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CalendarBackendServiceApplication.class, args);
-	}
-
-}
