--- src/test/java/calendar/backend/service/calendarbackendservice/TestConfiguration.java
@@ -1,15 +1,0 @@
-package calendar.backend.service.calendarbackendservice;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class TestConfiguration {
-
-    @Test
-    public void contextLoads() {
-    }
-}
